+++ .config/nvim/init.lua
@@ -1,0 +1,7 @@
+-- ========================================================================== --
+--                    NEOVIM CONFIGURATION ENTRY POINT (init.lua)              --
+-- ========================================================================== --
+
+require("config.options")
+require("config.keymaps")
+require("config.lazy")