+++ .config/nvim/lua/plugins/treesitter.lua
@@ -1,0 +1,19 @@
+return {
+  -- [ TREESITTER (Coloration syntaxique AST - Branche Master stable) ]
+  {
+    "nvim-treesitter/nvim-treesitter",
+    branch = "master",
+    build = ":TSUpdate",
+    config = function()
+      require("nvim-treesitter.configs").setup({
+        ensure_installed = { "lua", "vim", "vimdoc", "markdown", "markdown_inline", "javascript", "typescript", "html", "css", "bash" },
+        auto_install = true,
+        highlight = {
+          enable = true,
+          additional_vim_regex_highlighting = false,
+        },
+        indent = { enable = true },
+      })
+    end,
+  },
+}